[Groups UI] Properly handle groups on exam activity and implements RecyclerView to properly display groups. Starts to fix adapter

+++ app/src/test/java/fga/mds/gpp/trezentos/Controller/SortStudentsUtilUnitTest.java
@@ -49,7 +49,7 @@ public class SortStudentsUtilUnitTest{
             }
         });
 
-        Map<String, Double> mapOut = sortStudentsUtil.newMapStudents(list, 3, 10);
+        Map<String, Integer> mapOut = sortStudentsUtil.newMapStudents(list, 3, 10);
         Map<String, Double> expected = populateOrderedMapByGroupNumber();
 
         assertThat(mapOut, is(expected));
